Temp: Add files for in-tree testing of the ebuilds
The ./bin/test script tests the installation of 8 packages containing
only a single file. These packages only have an src_install phase, and
are cached as binpkgs for reuse in further invocations.

This allows me to consistently compare improvements, and share this
project with people to test on their own machines.

The goal of this small optimization project, is to improve the time it
takes to install binary packages. It currently takes way too long.

$ rm -rf .root && ./bin/test && hyperfine ./bin/test
Benchmark 1: ./bin/test
  Time (mean ± σ):     31.833 s ±  0.486 s    [User: 17.597 s, System: 8.048 s]
  Range (min … max):   30.919 s … 32.708 s    10 runs

PACKAGES PER SECOND (PPS): 0.25

(This project was motivated by the frustrating experience of waiting
multiple hours to install 1500 packages while creating a LiveCD)