Temp: Allow disabling the IPC daemon
Disabling the IPC daemon roughly *halves* the time ./bin/test takes on
my machine (with binpkgs). From 24 seconds to 12 seconds.

This indicates to me that there's a lot to gain from optimizing the IPC
mechanism.

These changes introduce a method by which to disable the IPC, allowing
me to compare the times with and without the IPC.

$ rm -rf .root && ./bin/test && PORTAGE_NOIPC=y hyperfine ./bin/test
Benchmark 1: ./bin/test
  Time (mean ± σ):     12.327 s ±  0.233 s    [User: 8.649 s, System: 5.047 s]
  Range (min … max):   12.010 s … 12.805 s    10 runs

PACKAGES PER SECOND (PPS): 0.65

(Is 1PPS possible? Find out in the next episode of whatever shit this is)