Optimization no. 2: Short-circuit ebuild-ipc exit
When an ebuild phase is empty, the only thing it needs to do is fall
through the entire script and report an exit status to portage.

Due to the IPC mechanism being enabled, this results in a call to
"ebuild-ipc exit".

I benchmarked the bin/misc-functions.sh script (it happened to be the
one driving the phase I was looking at), and the breakdown looks roughly
as follows:
	77.4ms - line 0 through right before the ebuild-ipc call
	117.7ms - the ebuild-ipc call itself
	- the script then inmediately calls exit

While there might be some room for improvement in the 77.4ms, it's hard
to deny that using over half of the script's runtime to report an exit
status is wasteful.

Benchmarking the bin/ebuild-ipc.py script itself, I could only conclude
that the vast majority of time was spent importing various modules, with
the most time being spent in "import portage". Making sure all the
python code is compiled and optimized through "python3 -m compileall -o2
./lib" didn't make a significant difference either.

That's why, I decided to short-circuit "ebuild-ipc exit" with the
simplest conceivable implementation in Python. One could argue that this
implementation isn't proper, as it doesn't handle locking, or checking
if the daemon is available, but: this ebuild-ipc call will never be
executed concurrently, and there isn't any reasonable thing ebuild.sh
can do if the IPC daemon is dead, besides exiting asap.

The new script executes in roughly 34ms, which is still bad, but much
faster, as can be seen in the benchmarks. Blame python startup times.

$ rm -rf .root && ./bin/test -q && hyperfine ./bin/test
Benchmark 1: ./bin/test
  Time (mean ± σ):     15.437 s ±  0.125 s    [User: 10.876 s, System: 5.943 s]
  Range (min … max):   15.270 s … 15.656 s    10 runs

PACKAGES PER SECOND (PPS): 0.52

And a current NOIPC benchmark for reference:

$ rm -rf .root && ./bin/test -q && PORTAGE_NOIPC=y hyperfine ./bin/test
Benchmark 1: ./bin/test
  Time (mean ± σ):     12.036 s ±  0.165 s    [User: 8.445 s, System: 4.925 s]
  Range (min … max):   11.867 s … 12.392 s    10 runs