Optimization no. 2.1: But wait, we can go deeper
This is an expansion of the same idea as last optimization: What if we
can bring down the runtime of the "ebuild-ipc exit" command?

The python interpreter's startup time is notoriously awful. Right now,
we're spending 34ms on this python script, about 30% of the total
runtime of misc-functions.sh.

The ebuild-ipc program can be rewritten in a different language. This
codebase currently makes it difficult to build and test tools written in
C, so bash is the only reasonable alternative.

Python's pickle format is a bit obtuse to interface with in bash, but
after fiddling a bit with the format I golfed something sensible,
using pickle format version 1. The "cat" program will exit as soon as it
gets a message + end of file from the IPC daemon, and the "dd" program
is used to ensure the message is sent in a single chunk, instead of being
buffered on the newlines.

The IPC system uses pickle, but it's only sending lists of strings (and
the exit status as an integer). It may make sense to instead use JSON,
or a custom format, to make it easier to interoperate with a different
programming language, if ebuild-ipc is ever to be reimplemented.

This new method executes in roughly 4ms, which is a vast improvement:

$ rm -rf .root && ./bin/test -q && hyperfine ./bin/test
Benchmark 1: ./bin/test
  Time (mean ± σ):     12.436 s ±  0.114 s    [User: 8.758 s, System: 5.141 s]
  Range (min … max):   12.266 s … 12.656 s    10 runs

PACKAGES PER SECOND (PPS): 0.64

For reference, a current NOIPC benchmark:

$ rm -rf .root && ./bin/test -q && PORTAGE_NOIPC=y hyperfine ./bin/test
Benchmark 1: ./bin/test
  Time (mean ± σ):     12.186 s ±  0.119 s    [User: 8.580 s, System: 4.948 s]
  Range (min … max):   12.047 s … 12.424 s    10 runs

The remaining extra runtime in the IPC-enabled version is likely the
setup time of the IPC daemon, and the 10ms exit delay. At this point I
feel the performance has been sufficiently matched, and there's bigger
fish to fry. I consider this a major win.