Optimization no. 2.2: Fix synchronization
Shell scripts don't guarnatee that the input file is opened before
spawning the backgrounded command.

By putting it in a pipeline, the shell will open the input fifo, block
because it has no writers, and then execute the remainder of the
pipeline, allowing us to send the command to the python program, and
unblock as soon as a full reply is received.